LAB-6/requirements.txt
Flask==0.12
pyjwt==1.4.2
Werkzeug==0.11.10
